Fix flow routing logic (fail early when flow rules block req), add tests
When access to a matching service was denied, tinyrange used to keep
searching for another matching service (which is futile because
ip+port combinations should be unique), and this led to confusing
diagnostics when a request was denied.

Fixed a flow rule logic bug that caused requests from instances with
no tags to always fail (caused by a spurious check).